integration tests: consolidate lxd/multipass fixtures for reuse
