+++ tests/integration/multipass/conftest.py
@@ -13,37 +13,16 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 """Fixtures for Multipass integration tests."""
-import os
-import pathlib
-import random
-import string
 import subprocess
-import sys
-import tempfile
 import time
 from contextlib import contextmanager
 
 import pytest
 
-from craft_providers import multipass
 
-
-@pytest.fixture()
-def home_tmp_path():
-    """Multipass doesn't have access to /tmp."""
-    with tempfile.TemporaryDirectory(
-        suffix=".tmp-pytest", dir=pathlib.Path.home()
-    ) as temp_dir:
-        yield pathlib.Path(temp_dir)
-
-
-def generate_instance_name():
-    return "itest-" + "".join(random.choices(string.ascii_uppercase, k=8))
-
-
-@pytest.fixture()
-def instance_name():
-    yield generate_instance_name()
+@pytest.fixture(autouse=True, scope="module")
+def installed_multipass_required(installed_multipass):
+    """All Multipass integration tests required multipass to be installed."""
 
 
 @contextmanager
@@ -101,46 +80,3 @@ def tmp_instance(
         capture_output=True,
         check=False,
     )
-
-
-@pytest.fixture(autouse=True, scope="module")
-def installed_multipass():
-    """Ensure multipass is installed, or skip the test if we cannot.
-
-    If the environment has CRAFT_PROVIDERS_TESTS_ENABLE_MULTIPASS_INSTALL=1,
-    force the installation of Multipass if uninstalled.
-    """
-    if multipass.is_installed():
-        return
-
-    if os.environ.get("CRAFT_PROVIDERS_TESTS_ENABLE_MULTIPASS_INSTALL") == "1":
-        multipass.install()
-    else:
-        pytest.skip("multipass not installed, skipped")
-
-
-@pytest.fixture
-def uninstalled_multipass():
-    """Uninstall Multipass prior to test, if environment allows it.
-
-    Environment may enable this fixture with:
-    CRAFT_PROVIDERS_TESTS_ENABLE_MULTIPASS_UNINSTALL=1
-    """
-    if not multipass.is_installed():
-        pytest.skip("multipass not installed, skipped")
-
-    if not os.environ.get("CRAFT_PROVIDERS_TESTS_ENABLE_MULTIPASS_UNINSTALL") == "1":
-        pytest.skip("not configured to uninstall multipass, skipped")
-
-    if sys.platform == "linux":
-        subprocess.run(["sudo", "snap", "remove", "multipass", "--purge"], check=True)
-    elif sys.platform == "darwin":
-        subprocess.run(["brew", "uninstall", "multipass"], check=True)
-    else:
-        pytest.skip("platform not supported to uninstall multipass, skipped")
-
-    yield
-
-    # Ensure it is installed after test.
-    if not multipass.is_installed():
-        multipass.install()
